lua: use preprocessor define to rename Lua main to unused_main
This way this library can have zero main functions defined.
If only the Lua main is defined, it will try to read from
stdin, which causes alert dialogs in a browser.
If I define a second main to override it, this causes an
error in a browser, and crashes the renderer process of
electron apps.