multiplayer/core: add APIs to get the multiplayer session ID.
This is the 5 digit code in the URL of the web version, used
by the websocket server to decide which players to connect a
player with.

The wxWidgets implementation doesn't need this, since its server
is only for a single session.

I think I implemented a normal socket client for the native
Android version too.